refactor(docs-retrieval): extract fetch-raw transport core from content-fetcher (#1115)
* refactor(docs-retrieval): extract fetch-raw transport core from content-fetcher

Move the network-facing transport core out of content-fetcher.ts into a
focused sibling module `content-fetcher/fetch-raw.ts` (PR 8 of the
content-fetcher modularization stack). content-fetcher.ts drops to ~578 LOC —
`fetchContent` is now a thin orchestrator over the transport + the loaders.

## What moved (verbatim, no logic change)

- `FetchError` + `FetchRawResult` types
- `enforceHttps` — the HTTPS-only security gate
- `generateUserFriendlyError` — FetchError → user-facing message mapping
- `tryUrlVariations` — ordered candidate loop (interactive-learning URLs)
- `tryGrafanaDocsContentLadder` — the content.json→unstyled.html ladder (PR 7)
- `fetchRawHtml` — the primary fetch, redirect trust re-validation, the manual
  3xx redirect branch, and error classification

## Barrel stability

All of this was module-private and only used internally by `fetchContent`,
which now imports `enforceHttps`, `fetchRawHtml`, and `generateUserFriendlyError`
back. No barrel change, no consumer touched. content-fetcher.ts sheds its now
unused imports: `DEFAULT_CONTENT_FETCH_TIMEOUT`, four security helpers
(`isAllowedContentUrl`, `isGrafanaDocsUrl`, `isLocalhostUrl`,
`isInteractiveLearningUrl`), and three url-utils helpers (`isJsonContentUrl`,
`generateInteractiveLearningVariations`, `getContentUrls`) — all now imported
by fetch-raw.ts instead. It keeps `parseUrlSafely` + `isTrustedFinalUrl`
(determineContentType / the entry trust gate) and `generateUrlId`
(wrapContentAsJsonGuide).

## Tests (+19, gap #1 — top priority)

The manual 3xx redirect branch was the most complex, most security-sensitive
untested code in the file. New `fetch-raw.test.ts` pins it directly:
- follows a relative same-origin redirect to a trusted target
- does NOT follow an absolute (non-`/`) Location — surfaces the redirect error,
  target never fetched
- 3xx with no Location header
- redirect target fetch throws / returns empty (error retained)
Plus trust re-validation (ok response redirected to an untrusted final URL),
error classification (404→not-found, 500→server-error, abort→timeout,
Failed-to-fetch→network), `enforceHttps` (https ok / http rejected / malformed),
and the full `generateUserFriendlyError` mapping table.

## Reviewer note

The cross-tier surface is unchanged — fetch-raw.ts is Tier 2 and imports only
downward (types, constants, security, utils) plus its sibling url-utils, so
`architecture.test.ts` stays green with no new allowlist entry.


* refactor(docs-retrieval): extract package-content + resolver singleton (final) (#1116)

Final PR of the content-fetcher modularization stack. Moves the package-content
integration — including the module-level `_packageResolver` singleton — out of
content-fetcher.ts into `content-fetcher/package-content.ts`. content-fetcher.ts
is now a ~290-line orchestrator: `fetchContent` + `wrapContentAsJsonGuide` +
the small URL/content-type helpers, composing the extracted loaders and
transport.

## What moved (verbatim, no logic change)

- `_packageResolver` singleton + `setPackageResolver`
- `derivePathSlug`, `buildMilestoneWebsiteUrl`
- `resolvePackageMilestones`, `resolvePackageNavLinks`
- `isPathManifest`, `getManifestMilestoneIds`
- `fetchPackageContent` (incl. the `Promise.all` interleaving of content fetch /
  milestone resolution / baseUrl resolution — preserved exactly), `fetchPackageById`

## Breaking the cycle (reviewer gotcha)

`package-content.ts` imports `fetchContent` from `../content-fetcher`, and
`fetchPackageContent` calls `injectJourneyExtrasIntoJsonGuide` (imported from
`./cover-page`). The dependency is **one-directional**: content-fetcher.ts no
longer references any package code, so there is no import cycle. To keep that
true, the barrel's package re-export block now pulls these names from
`./content-fetcher/package-content` instead of `./content-fetcher` — the
**public surface is unchanged** (same six names exported from
`docs-retrieval`), only the internal source path moved.

content-fetcher.ts sheds five now-unused imports (`PackageResolver`,
`ResolvedNavLink`, `getPackageRenderType`, `Milestone`, `LearningJourneyMetadata`).

## Tests

The existing `package-content.test.ts` (36 tests) is the invariant safety net —
its assertions are untouched. Two moved-seam fixes only: the package-fn import
repoints to `./content-fetcher/package-content` (with `fetchContent` still from
`./content-fetcher`). Added happy-path coverage for the two previously-untested
seams: `resolvePackageNavLinks` (resolution, title fallback, skip-unresolvable,
empty/no-resolver guards) and `buildMilestoneWebsiteUrl` via
`resolvePackageMilestones` with a path slug (prefix match / no match / no slug).

A backup branch `refactor-content-fetcher-pre-pr9` was cut before this change.


---------
